# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API para radioxenter",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:4000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/convenios/{cod_convenio}": {
      "get": {
        "tags": [
          "Convenios"
        ],
        "description": "Endpoint para obtener un convenio",
        "parameters": [
          {
            "name": "cod_convenio",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/convenios": {
      "get": {
        "tags": [
          "Convenios"
        ],
        "description": "Endpoint para obtener convenios",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Convenios"
        ],
        "description": "Endpoint para registrar un convenio.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Convenio"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Convenios"
        ],
        "description": "Endpoint para editar un convenio.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Convenio"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/empleados/{cod_empleado}": {
      "get": {
        "tags": [
          "Empleados"
        ],
        "description": "Endpoint para obtener un empleado",
        "parameters": [
          {
            "name": "cod_empleado",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/empleados": {
      "get": {
        "tags": [
          "Empleados"
        ],
        "description": "Endpoint para obtener empleados",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Empleados"
        ],
        "description": "Endpoint para editar un empleado.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Empleado"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/entidades/{cod_entidad}/convenios": {
      "get": {
        "tags": [
          "Entidades"
        ],
        "description": "Endpoint para obtener todos los convenios de una entidad",
        "parameters": [
          {
            "name": "cod_entidad",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Entidades"
        ],
        "description": "Endpoint para editar los convenios de una entidad.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EditarConvenioEntidad"
            }
          },
          {
            "name": "cod_entidad",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/entidades/{cod_entidad}/convenios/{cod_servicio}": {
      "put": {
        "tags": [
          "Entidades"
        ],
        "description": "Endpoint para editar un servicio de un convenio con una entidad.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ConvenioUpdate"
            }
          },
          {
            "name": "cod_entidad",
            "in": "path",
            "required": true
          },
          {
            "name": "cod_servicio",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/entidadDoctor/{cod_entidad_doctor}": {
      "get": {
        "tags": [
          "EntidadDoctor - DEPRECATED"
        ],
        "description": "Endpoint para obtener una entidad doctor",
        "parameters": [
          {
            "name": "cod_entidad_doctor",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/entidadDoctor": {
      "get": {
        "tags": [
          "EntidadDoctor - DEPRECATED"
        ],
        "description": "Endpoint para obtener entidad doctores",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "EntidadDoctor - DEPRECATED"
        ],
        "description": "Endpoint para crear una entidad doctor.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EntidadDoctor"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "EntidadDoctor - DEPRECATED"
        ],
        "description": "Endpoint para editar una entidad doctor.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EntidadDoctor"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Empleado": {
      "type": "object",
      "properties": {
        "cod_empleado": {
          "type": "number",
          "example": 0
        },
        "nombres_empleado": {
          "type": "string",
          "example": ""
        },
        "apellidos_empleado": {
          "type": "string",
          "example": ""
        },
        "documento_empleado": {
          "type": "number",
          "example": 0
        },
        "direccion_empleado": {
          "type": "string",
          "example": ""
        },
        "fnacimiento_empleado": {
          "type": "string",
          "example": "dd/MM/YYYYY"
        },
        "telefono_empleado": {
          "type": "string",
          "example": ""
        },
        "correo_empleado": {
          "type": "string",
          "example": "email@email.com"
        },
        "contrasenia_empleado": {
          "type": "string",
          "example": ""
        },
        "usuario_empleado": {
          "type": "string",
          "example": ""
        },
        "cod_tipo_empleado": {
          "type": "number",
          "example": 0
        },
        "cod_tipo_documento": {
          "type": "number",
          "example": 0
        }
      },
      "xml": {
        "name": "Empleado"
      }
    },
...
    "EditarConvenioEntidad": {
      "type": "object",
      "properties": {
        "servicios_convenio": {
          "type": "array",
          "items": {
            "type": "number"
          }
        },
        "valores_servicios": {
          "type": "array",
          "items": {
            "type": "number"
          }
        },
        "fecha_inicial_convenio": {
          "type": "string",
          "example": "dd/MM/YYYYY"
        },
        "fecha_final_convenio": {
          "type": "string",
          "example": "dd/MM/YYYYY"
        }
      },
      "xml": {
        "name": "EditarConvenioEntidad"
      }
    },
...
    "Convenio": {
      "type": "object",
      "properties": {
        "valor_servicio": {
          "type": "number",
          "example": 0
        },
        "fecha_inicial_convenio": {
          "type": "string",
          "example": "yyyy-MM-dd"
        },
        "fecha_final_convenio": {
          "type": "string",
          "example": "yyyy-MM-dd"
        },
        "cod_entidad": {
          "type": "number",
          "example": 0
        },
        "cod_servicio": {
          "type": "number",
          "example": 0
        }
      },
      "xml": {
        "name": "Convenio"
      }
    },
    "ConvenioUpdate": {
      "type": "object",
      "properties": {
        "valor_servicio": {
          "type": "number",
          "example": 0
        },
        "fecha_inicial_convenio": {
          "type": "string",
          "example": "yyyy-MM-dd"
        },
        "fecha_final_convenio": {
          "type": "string",
          "example": "yyyy-MM-dd"
        }
      },
      "xml": {
        "name": "ConvenioUpdate"
      }
    },
    "EntidadDoctor": {
      "type": "object",
      "properties": {
        "cod_entidad": {
          "type": "number",
          "example": 0
        },
        "cod_doctor": {
          "type": "number",
          "example": 0
        }
      },
      "xml": {
        "name": "EntidadDoctor"
      }
    },
...
  }
}